Rename ApiModel → OpenAPI and ModelBuilder → OpenAPI::Builder
Reorganize namespace for better clarity and alignment with OpenAPI spec:

Module renames:
- GrapeSwagger::ApiModel::Spec → GrapeSwagger::OpenAPI::Document
- GrapeSwagger::ModelBuilder::DirectSpecBuilder → GrapeSwagger::OpenAPI::Builder::FromRoutes
- GrapeSwagger::ModelBuilder::SpecBuilder → GrapeSwagger::OpenAPI::Builder::FromHash
- All ApiModel classes moved to OpenAPI namespace

Directory structure:
- lib/grape-swagger/api_model/ → lib/grape-swagger/openapi/
- lib/grape-swagger/model_builder/ → lib/grape-swagger/openapi/builder/

Updated references in:
- lib/grape-swagger.rb
- lib/grape-swagger/doc_methods.rb
- lib/grape-swagger/exporter/*.rb
- spec/openapi_v3/*.rb
- docs/*.md
