+++ lib/grape-swagger/openapi/tag.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module GrapeSwagger
-  module ApiModel
+  module OpenAPI
     # Tag definition for grouping operations.
     class Tag
       attr_accessor :name, :description, :external_docs, :extensions
